Split UIUnitTest into two libs
