+++ Lib/Sources/UIUnitTest/AccessibilityAuditType.swift
@@ -1,11 +1,5 @@
-//
-//  File.swift
-//  
-//
-//  Created by Bruno Mazzo on 7/5/2024.
-//
-
 import Foundation
+import UIUnitTestAPI
 
 public struct AccessibilityAuditType: RawRepresentable, OptionSet, Codable, Sendable {
     
@@ -34,7 +28,7 @@ public struct AccessibilityAuditType: RawRepresentable, OptionSet, Codable, Send
     public static let all                          = AccessibilityAuditType(rawValue: ~0)
 }
 
-public struct AccessibilityAuditIssue: Codable, Sendable {
+public struct AccessibilityAuditIssue: Sendable {
     
     /// The element associated with the issue.
     public var element: Element?
@@ -47,58 +41,16 @@ public struct AccessibilityAuditIssue: Codable, Sendable {
     
     /// The type of audit which generated the issue.
     public var auditType: AccessibilityAuditType
-    
-    public init(element: UUID? = nil, compactDescription: String, detailedDescription: String, auditType: AccessibilityAuditType) {
-        if let element = element {
+   
+    public init(data: AccessibilityAuditIssueData) {
+        if let element = data.element {
             self.element = Element(serverId: element)
         } else {
             self.element = nil
         }
         
-        self.compactDescription = compactDescription
-        self.detailedDescription = detailedDescription
-        self.auditType = auditType
-    }
-    
-    enum CodingKeys: CodingKey {
-        case element
-        case compactDescription
-        case detailedDescription
-        case auditType
-    }
-    
-    public init(from decoder: any Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        
-        if let element = try container.decodeIfPresent(UUID.self, forKey: .element) {
-            self.element = Element(serverId: element)
-        }
-        
-        self.compactDescription = try container.decode(String.self, forKey: .compactDescription)
-        self.detailedDescription = try container.decode(String.self, forKey: .detailedDescription)
-        self.auditType = try container.decode(AccessibilityAuditType.self, forKey: .auditType)
-    }
-    
-    public func encode(to encoder: any Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        
-        try container.encodeIfPresent(element?.serverId, forKey: .element)
-        try container.encode(compactDescription, forKey: .compactDescription)
-        try container.encode(detailedDescription, forKey: .detailedDescription)
-        try container.encode(auditType, forKey: .auditType)
-    }
-}
-
-public struct AccessibilityAuditRequest: Codable, Sendable {
-    public var serverId: UUID
-    public var accessibilityAuditType: AccessibilityAuditType
-}
-
-
-public struct AccessibilityAuditResponse: Codable, Sendable {
-    public var issues: [AccessibilityAuditIssue]
-    
-    public init(issues: [AccessibilityAuditIssue]) {
-        self.issues = issues
+        self.compactDescription = data.compactDescription
+        self.detailedDescription = data.detailedDescription
+        self.auditType = AccessibilityAuditType(rawValue: data.auditType)
     }
 }
