+++ Lib/Sources/UIUnitTest/Element/Element+Swipe.swift
@@ -1,12 +1,16 @@
 import Foundation
+import UIUnitTestAPI
 
-public enum SwipeDirection: Codable, Sendable {
-    case up, down, left, right
+public enum SwipeDirection: Int, Codable, Sendable {
+    case up = 1
+    case down = 2
+    case left = 3
+    case right = 4
 }
 
 extension Element {
     public func swipe(direction: SwipeDirection, velocity: GestureVelocity = .default) async throws {
-        let swipeRequest = SwipeRequest(serverId: serverId, direction: direction, velocity: velocity)
+        let swipeRequest = SwipeRequest(serverId: serverId, direction: direction.rawValue, velocity: velocity)
         
         let _: Bool = try await callServer(path: "swipe", request: swipeRequest)
     }
@@ -55,40 +59,7 @@ extension Element {
     }
 }
 
-public struct SwipeRequest: Codable, Sendable {
-    
-    public var serverId: UUID
-    public var swipeDirection: SwipeDirection
-    public var velocity: GestureVelocity
-    
-    public init(serverId: UUID, direction: SwipeDirection, velocity: GestureVelocity) {
-        self.serverId = serverId
-        self.swipeDirection = direction
-        self.velocity = velocity
-    }
-}
 
-public enum GestureVelocity: Hashable, Equatable, @unchecked Sendable, Codable {
-    case `default`, slow, fast
-    case custom(CGFloat)
-    
-    public init(_ value: CGFloat) {
-        self = .custom(value)
-    }
-}
-
-public struct GestureVelocity2 : Hashable, Equatable, RawRepresentable, @unchecked Sendable, Codable {
-    
-    public let rawValue: CGFloat
-    
-    public init(_ rawValue: CGFloat) {
-        self.rawValue = rawValue
-    }
-    
-    public init(rawValue: CGFloat) {
-        self.rawValue = rawValue
-    }
-}
 
 extension GestureVelocity : ExpressibleByIntegerLiteral {
